Transform KTX2Loader and WorkerPool
