+++ package.json
@@ -59,9 +59,9 @@
     "@storybook/html": "^6.4.14",
     "@storybook/manager-webpack5": "^6.4.14",
     "@storybook/preset-typescript": "^3.0.0",
-    "@types/draco3d": "^1.4.0",
+    "@types/draco3d": "^1.4.2",
     "@types/offscreencanvas": "^2019.6.4",
-    "@types/three": "^0.139.0",
+    "@types/three": "^0.146.0",
     "@typescript-eslint/eslint-plugin": "^4.28.0",
     "@typescript-eslint/parser": "^4.28.0",
     "babel-loader": "^8.2.3",
@@ -90,7 +90,7 @@
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-size-snapshot": "^0.12.0",
     "rollup-plugin-terser": "^7.0.2",
-    "three": "^0.139.2",
+    "three": "0.146.0",
     "tslib": "^2.2.0",
     "typescript": "^4.2.4",
     "webpack": "^5.65.0"
